backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
supabase>=2.16.0
httpx>=0.28.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.25
alembic==1.13.1
redis==5.0.1
celery==5.3.4
pytest==7.4.4
pytest-asyncio==0.21.1
python-dateutil==2.8.2
numpy>=1.26.4
pandas>=2.2.0
scipy>=1.13.0
scikit-learn>=1.4.0
openai==1.7.1
anthropic>=0.25.0
google-generativeai==0.3.2
groq==0.4.2
tiktoken==0.5.2
langchain==0.1.0
pinecone-client>=2.2.0
# chromadb>=0.5.0  # Commented out due to Python 3.13 compatibility issues
# tavily-python==0.3.0  # Commented out due to tiktoken dependency conflict
firecrawl-py==0.0.16
beautifulsoup4==4.12.2
lxml==5.1.0
aiofiles==23.2.1
aiohttp==3.9.1
certifi>=2024.2.2
structlog==24.1.0
prometheus-client==0.19.0
matplotlib==3.10.3
seaborn==0.13.2
stripe==7.9.0
yfinance==0.2.28
python-pptx==1.0.2
reportlab==4.2.5
playwright==1.40.0